requirements.txt
hydra-core
hydra-colorlog
torch
torchvision
cupy-cuda102
pytorch3d
pytorch-lightning
numpy
matplotlib
opencv-python
lpips
tqdm
wandb
ConfigSpace
ray[tune]
